build/libvfn: Enable support for external libvfn
Now Meson is able to find a system-installed libvfn and compile with
that. Thus, the libvfn subproject has been removed since it is no longer
needed.

This relies on Meson actually being able to find the system-installed
libvfn. If 'XNVME_BE_LINUX_VFIO_ENABLED' is true and it cannot find libvfn,
the build will fail. This is to avoid Meson accidentally overwriting the system-installed
libvfn as in the example below:

* libvfn is installed in /usr/local/lib
* /usr/local/lib/pkgconfig is not part of the PKG_CONFIG_PATH
* During installation of xNVMe Meson cannot find libvfn
* Meson installs xNVMe and the libvfn subproject to
  /usr/local/lib

Signed-off-by: Karl Bonde Torp <k.torp@samsung.com>